+++ R/qualpal.R
@@ -40,8 +40,8 @@
 #'     \item{A \code{\link{list}} with the following \emph{named} vectors,
 #'       each of length two, giving a range for each item.}{
 #'       \describe{
-#'         \item{\code{h}}{Hue, in the range [0, 1]}
-#'         \item{\code{s}}{Saturation, in the range [-360, 360]}
+#'         \item{\code{h}}{Hue, in the range [-360, 360]}
+#'         \item{\code{s}}{Saturation, in the range [0, 1]}
 #'         \item{\code{l}}{Lightness, in the range [0, 1]}
 #'       }
 #'     }