+++ NEWS.md
@@ -4,6 +4,7 @@
 * `qualpal()` gains an argument, `n_threads`, for specifying the number
 of threads to use when computing the distance matrix between colors.
 * C++ functions call namespaces explicitly using `::`.
+* Documentation for hue and saturation in `qualpal()` has been fixed. (#2)
 
 # qualpalr 0.4.2